requirements.txt
paddlepaddle==2.5.1
paddleocr>=2.0.1
numpy
pyautogui
tk
Pillow
opencv-python
deepl